cdr_stats/billing_requirements.txt
django-threaded-multihost==1.4-0
hg+https://bitbucket.org/jespern/django-piston
